==> Cleaning/1. Cleaning Restaurants data.ipynb <==
# Notebook 1:  Cleaning of Spanish Michelin Star Restaurants' database

The used database has been created bases on [Michelin Guide website](https://guide.michelin.com/es/es).

## 1. Importing libraries

import pandas as pd
import numpy as np
import sys
sys.path.insert(0, '../src')
import unittest
from call import *

## 2. Importing database and data exploration¶


restaurants = pd.read_csv("../data/restaurants.csv", encoding = "ISO-8859-1",sep = (";"), engine='python', error_bad_lines=False, warn_bad_lines=False)

restaurants.shape

restaurants.head()

restaurants.columns

restaurants.dtypes

## 3. Cleaning

### 3.1. Checking for nulls

resturants = restaurants.drop(restaurants[restaurants.CCAA.isnull()].index, inplace = True)

restaurants.shape

### 3.2. Renaming columns

restaurants = restaurants.rename(columns={'International_cooking': 'international_cooking'})
restaurants = restaurants.rename(columns={'CCAA': 'region'})
restaurants = restaurants.rename(columns={'restaurant_name': 'restaurant'})

restaurants.shape

### 3.3. Replacing values

restaurants.loc[restaurants["region"] == "Catalua", "region"] = "Cataluna"
restaurants.loc[restaurants["province"] == "La Corua", "province"] = "La Coruna"
restaurants.loc[restaurants["city"] == "La Corua", "city"] = "La Coruna"

### 3.4. Spliting coordenates column into latitude and longitude

coordenates = restaurants["coordenates"].str.split(",", n = 1, expand = True)

restaurants["latitude"]= coordenates[0]
restaurants["longitude"]= coordenates[1]

### 3.5. Reordering columns

restaurants = restaurants.reindex(columns = ['restaurant','michelin_stars','city','province','region','coast_province','coordenates','latitude','longitude','price_menu_1','price_menu_2','price_average'])

## 4. Exporting cleaned data

restaurants = restaurants.reset_index(drop=True)

restaurants.shape

restaurants.head()

restaurants.to_csv("../output/restaurants.csv", index = False)

restaurants_region = restaurants.groupby(['region']).count()['restaurant']
restaurants_region.head()

restaurants_region.to_csv("../output/restaurants_region.csv", index = False)
