==> Cleaning/2. Cleaning employment data.ipynb <==
# In this second notebook the cleaning of Spanish Q4 2020 employment rates database takes place.

The database was obtained from [INE website](https://www.ine.es/).

## 1. Importing libraries that are used.

import pandas as pd
import numpy as np
import sys
sys.path.insert(0, '../src')
import unittest
from call import *

## 2. Importing database and data exploration

employment = pd.read_csv("../data/empleo.csv", encoding = "ISO-8859-1",sep = (";"), engine='python', error_bad_lines=False, warn_bad_lines=False)

employment.shape

employment.head()

employment.columns

employment.dtypes

## 3. Cleaning database

### 3.1. Renaming information content of an specific column.

employment = employment.rename(columns={'Total': 'Employment rate'})

### 3.2. Spliting information conent in column named "Provincias".

provincias = employment["Provincias"].str.split(" ", n = 1, expand = True)

employment["Provincias"] = provincias[1]

employment["Provincias"].unique()

### 3.3. Replacing values.

#### 3.3.1. Replacing vowels

# Function imported from call.py
employment["Provincias"] = employment["Provincias"].apply(normalize)

#### 3.3.2. Replacing the names of the provinces with the names that appear in the restaurants' database

employment.loc[employment["Provincias"] == "Alicante/Alacant", "Provincias"] = "Alicante"
employment.loc[employment["Provincias"] == "Coruña, A", "Provincias"] = "La Coruna"
employment.loc[employment["Provincias"] == "Rioja, La", "Provincias"] = "La Rioja"
employment.loc[employment["Provincias"] == "Bizkaia", "Provincias"] = "Vizcaya"
employment.loc[employment["Provincias"] == "Araba/Alava", "Provincias"] = "Alava"
employment.loc[employment["Provincias"] == "Castellon/Castello", "Provincias"] = "Castellon"
employment.loc[employment["Provincias"] == "Lleida", "Provincias"] = "Lerida"
employment.loc[employment["Provincias"] == "Gipuzkoa", "Provincias"] = "Guipuzcua"
employment.loc[employment["Provincias"] == "Girona", "Provincias"] = "Gerona"
employment.loc[employment["Provincias"] == "Valencia/València", "Provincias"] = "Valencia"
employment.loc[employment["Provincias"] == "Balears, Illes", "Provincias"] = "Islas Baleares"
employment.loc[employment["Provincias"] == "Palmas, Las", "Provincias"] = "Las Palmas"

employment.Provincias.unique()

employment.shape

### 3.4. Drop unnecessary column

employment = employment.drop(['Tasas','Periodo','Sexo'], axis=1)

employment.head()

## 4. Exporting cleaning data

employment = employment.reset_index(drop = True)

employment.head()

employment.to_csv("../output/employment.csv", index = False)
