==> Cleaning/3. Cleaning unemployment data.ipynb <==
# In this third cleaning notebook takes place the cleaning of Spanish Q4 2020 unemployment rates database.

The database was obtained from [INE website](https://www.ine.es/).

## 1. Importing libraries that are used.

import pandas as pd
import numpy as np
import sys
sys.path.insert(0, '../src')
import unittest
from call import *

## 2. Importing database and data exploration

unemployment = pd.read_csv("../data/paro.csv", encoding = "ISO-8859-1",sep = (";"), engine='python', error_bad_lines=False, warn_bad_lines=False)

unemployment.shape

unemployment.head()

unemployment.columns

unemployment.dtypes

## 3. Cleaning database

### 3.1. Renaming information content of an specific column.

unemployment = unemployment.rename(columns={'Total': 'Unemployment rate'})

### 3.2. Spliting information conent in column named "Provincias".

provincias = unemployment["Provincias"].str.split(" ", n = 1, expand = True)

unemployment["Provincias"] = provincias[1]

unemployment["Provincias"].unique()

### 3.3. Replacing values.

#### 3.3.1. Replacing vowels

unemployment["Provincias"] = unemployment["Provincias"].apply(normalize)

#### 3.3.2. Replacing the names of the provinces with the names that appear in the restaurants' database

unemployment.loc[unemployment["Provincias"] == "Alicante/Alacant", "Provincias"] = "Alicante"
unemployment.loc[unemployment["Provincias"] == "Coruña, A", "Provincias"] = "La Coruna"
unemployment.loc[unemployment["Provincias"] == "Rioja, La", "Provincias"] = "La Rioja"
unemployment.loc[unemployment["Provincias"] == "Bizkaia", "Provincias"] = "Vizcaya"
unemployment.loc[unemployment["Provincias"] == "Araba/Alava", "Provincias"] = "Alava"
unemployment.loc[unemployment["Provincias"] == "Castellon/Castello", "Provincias"] = "Castellon"
unemployment.loc[unemployment["Provincias"] == "Lleida", "Provincias"] = "Lerida"
unemployment.loc[unemployment["Provincias"] == "Gipuzkoa", "Provincias"] = "Guipuzcua"
unemployment.loc[unemployment["Provincias"] == "Girona", "Provincias"] = "Gerona"
unemployment.loc[unemployment["Provincias"] == "Valencia/València", "Provincias"] = "Valencia"
unemployment.loc[unemployment["Provincias"] == "Balears, Illes", "Provincias"] = "Islas Baleares"
unemployment.loc[unemployment["Provincias"] == "Palmas, Las", "Provincias"] = "Las Palmas"

unemployment.Provincias.unique()

unemployment.shape

### 3.4. Drop unnecessary column

unemployment = unemployment.drop(['Tasas','Periodo','Sexo'], axis=1)

unemployment.head()

## 4. Exporting cleaning data

unemployment = unemployment.reset_index(drop = True)

unemployment.head()

unemployment.to_csv("../output/unemployment.csv", index = False)
