==> Cleaning/4. Cleaning tourists database.ipynb <==
# In this notebook, I proceed with the cleaning Spanish tourism during 2019 database.

The database was obtained from [INE website](https://www.ine.es/).

## 1. Importing libraries

import pandas as pd
import numpy as np
import sys
sys.path.insert(0, '../src')
import unittest
from call import *

## 2. Importing database and data exploration

tourists2019 = pd.read_csv("../data/tourists2019.csv", encoding = "ISO-8859-1",sep = (";"), engine='python', error_bad_lines=False, warn_bad_lines=False)

tourists2019.shape

tourists2019.head()

tourists2019.columns

tourists2019.dtypes

## 3. Cleaning database

### 3.1. Renaming columns

tourists2019 = tourists2019.rename(columns={'Comunidades autónomas': 'region'})

tourists2019

### 3.2. Delete worthless information

region = tourists2019["region"].str.split(" ", n = 1, expand = True)

tourists2019["region"] = region[1]

tourists2019["region"].unique()

### 3.3. Replacing column values

tourists2019.loc[tourists2019["region"] == "Andaluc\x92a", "region"] = "Andalucia"
tourists2019.loc[tourists2019["region"] == "Catalu\x96a", "region"] = "Cataluna"
tourists2019.loc[tourists2019["region"] == "Otras Comunidades Autónomas", "region"] = "Other regions"

tourists2019.shape

### 3.4. Converting Total values from object to float

tourists2019['Total'] = tourists2019['Total'].apply(lambda x: convert_str(x))

tourists2019.groupby(['region']).sum()['Total']

### 3.5. Grouping by region value

tourists2019 = (tourists2019.groupby(['region']).agg({"Total":"sum"}).reset_index())

tourists2019

## 4. Exporting cleaned data

tourists2019 = tourists2019.reset_index(drop = True)

tourists2019.head()

tourists2019.to_csv("../output/tourists2019.csv", index = False)
