==> Cleaning/5. Cleaning provinces population database.ipynb <==
# In this fifth  notebook, I proceed with the cleaning of Spanish 2020 population database.

The database was obtained from [INE website](https://www.ine.es/).

## 1. Importing libraries that are used.

import pandas as pd
import numpy as np
import sys
sys.path.insert(0, '../src')
import unittest
from call import *

## 2. Importing database and data exploration

population2020 = pd.read_csv("../data/population.csv", encoding = "ISO-8859-1",sep = (";"), engine='python', error_bad_lines=False, warn_bad_lines=False)

population2020.shape

population2020.head()

population2020.columns

population2020.dtypes

## 3. Cleaning data

### 3.1. Replacing values

population2020.loc[population2020["Provincias"] == "Alicante/Alacant", "Provincias"] = "Alicante"
population2020.loc[population2020["Provincias"] == "Coruña, A", "Provincias"] = "La Coruna"
population2020.loc[population2020["Provincias"] == "Rioja, La", "Provincias"] = "La Rioja"
population2020.loc[population2020["Provincias"] == "Bizkaia", "Provincias"] = "Vizcaya"
population2020.loc[population2020["Provincias"] == "Araba/Alava", "Provincias"] = "Alava"
population2020.loc[population2020["Provincias"] == "Castellon/Castello", "Provincias"] = "Castellon"
population2020.loc[population2020["Provincias"] == "Lleida", "Provincias"] = "Lerida"
population2020.loc[population2020["Provincias"] == "Gipuzkoa", "Provincias"] = "Guipuzcua"
population2020.loc[population2020["Provincias"] == "Girona", "Provincias"] = "Gerona"
population2020.loc[population2020["Provincias"] == "Palmas, Las", "Provincias"] = "Las Palmas"

population2020.Provincias.unique()

### 3.2. Renaming column

population2020 = population2020.rename(columns={'Total': 'Population'})

### 3.3. Converting values from object to float

population2020['Population'] = population2020['Population'].apply(lambda x: convert_str(x))

population2020.head()

## 4. Joining databases

### 4.1. Uploading dabases we want to join

employment = pd.read_csv("../output/employment.csv", encoding = "ISO-8859-1")
unemployment = pd.read_csv("../output/unemployment.csv", encoding = "ISO-8859-1")

### 4.2. Concatenating dataframes

province_data = pd.concat([population2020,employment,unemployment], axis=1)
province_data[:7]

### 4.3. Deleting duplicated columns

province_data = province_data.loc[:,~province_data.columns.duplicated(keep = 'first')]

province_data.head()

province_data.dtypes

## 5. Exporting joined data.

province_data.to_csv("../output/province_data.csv", index = False)
